logind: explicitly find session when socket activated
Right now, the logind support in Xorg only supports
being run from within a session.  That prevents it
from being dbus activated as part of the user session.

This commit changes logind to search through all the available
sessions of the system and try to find an appropriate session
to use when socket activated.

Acked-by: Peter Hutterer <peter.hutterer@who-t.net>