logind: Deactivate server when session goes away
While the X server's process lifetime is tied to the lifetime of a
particular user, when the X server is active, it is also associated
with a particular session of the user.  If that session closes, the
X server should reset and wait for the next session to come along, or
wait to be killed because the user has gone.

This commit makes it watch the session's state and deactivate/reactivate
as appropriate.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>