xfree86: detach from controlling terminal when closing console
Right now the X server keeps control of the tty associated with the
VT it's using even after the tty is closed.

In order for the X server to support changing which VT it uses at run
time we need to make sure it's not controlling an old tty when taking
on a new one.

This commit makes sure that when the console is closed, the process
compoletely cedes control of the terminal.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>