dix: add new dispatchMode for tracking how "up" the server is
This commit provides a way for the X server to be
"deactivated" where by it sits idle with no clients,
just waiting to be reset (and reactivated) or terminated.

This will be useful for supporting a socket activated
X server which is tied to the lifetime of the user instead of
a particular session.