os: don't require display number when socket activated
If the X server is socket activated, there's no opportunity
for clients and server to negotiate a display number.

They instead communicate via a well known socket name specific
to the user. The socket is already open before the X server starts and is
handed to the X server during initialization.

Since communication is already established without a display number,
this commit makes sure that a display number isn't required if
the server is socket activated.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>