xfree: probe devices again when reactivating
Right now we only probe graphics devices during
the initial bring up of the X server.

In the case of the logind codepath, this probing
process is where the drm fd is requested.

This commit makes sure we reprobe when reactivating,
so that logind can give us a fresh drm fd associated
with our new session.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>